+++ src/fhir/timeline-event.ts
@@ -51,7 +51,12 @@ export function useTimelineEvents() {
     );
     // Disabling because including queries will cause infinite loop
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [queries.map((q) => q.data)]);
+  }, [
+    patientEncountersQuery.data,
+    diagnosticReportQuery.data,
+    medicationRequestCommon.data,
+    medicationDispenseCommon.data,
+  ]);
 
   const isLoading = some(queries, "isLoading");
   const isFetching = some(queries, "isFetching");
